[feat #175] replace all buttons with Button component
